Refactor traffic light controller
Remove the while loop so that the simulation can
be run for an indefinite amount of time. Now, the user of
the controller needs to explicitly call the stop method to
stop the simulation and hence clear any timeouts. This is
also better for performance, as the timers are created
as needed instead of upfront.